refactor(sdk): Move compile_graph() to bento.graph.compile module to prevent recursive imports
